+++ NEWS.md
@@ -13,6 +13,8 @@
 
 * New `--list --json`.
 
+* Very basic mutation of `Result<_, _>` to return `Ok(Default::default())`.
+
 ## 0.0.2
 
 * Functions that should not be mutated can be marked with `#[mutants::skip]`